# Simplify BitReader (~5-10% faster)