--- a/parquet/src/util/bit_util.rs
+++ b/parquet/src/util/bit_util.rs
@@ -330,32 +330,23 @@ pub struct BitReader {
 
     // Current bit offset in `buffered_values`
     bit_offset: usize,
-
-    // Total number of bytes in `buffer`
-    total_bytes: usize,
 }
 
 /// Utility class to read bit/byte stream. This class can read bits or bytes that are
 /// either byte aligned or not.
 impl BitReader {
     pub fn new(buffer: ByteBufferPtr) -> Self {
-        let total_bytes = buffer.len();
-        let num_bytes = cmp::min(8, total_bytes);
-        let buffered_values = read_num_bytes!(u64, num_bytes, buffer.as_ref());
         BitReader {
             buffer,
-            buffered_values,
+            buffered_values: 0,
             byte_offset: 0,
             bit_offset: 0,
-            total_bytes,
         }
     }
 
     pub fn reset(&mut self, buffer: ByteBufferPtr) {
         self.buffer = buffer;
-        self.total_bytes = self.buffer.len();
-        let num_bytes = cmp::min(8, self.total_bytes);
-        self.buffered_values = read_num_bytes!(u64, num_bytes, self.buffer.as_ref());
+        self.buffered_values = 0;
         self.byte_offset = 0;
         self.bit_offset = 0;
     }
@@ -373,8 +364,13 @@ impl BitReader {
         assert!(num_bits <= 64);
         assert!(num_bits <= size_of::<T>() * 8);
 
-        if self.byte_offset * 8 + self.bit_offset + num_bits > self.total_bytes * 8 {
+        if self.byte_offset * 8 + self.bit_offset + num_bits > self.buffer.len() * 8 {
             return None;
+        }
+
+        // Only need to populate buffer if byte aligned
+        if self.bit_offset == 0 {
+            self.load_buffer()
         }
 
         let mut v = trailing_bits(self.buffered_values, self.bit_offset + num_bits)
@@ -385,9 +381,12 @@ impl BitReader {
             self.byte_offset += 8;
             self.bit_offset -= 64;
 
-            self.reload_buffer_values();
-            v |= trailing_bits(self.buffered_values, self.bit_offset)
-                .wrapping_shl((num_bits - self.bit_offset) as u32);
+            if self.bit_offset != 0 {
+                self.load_buffer();
+
+                v |= trailing_bits(self.buffered_values, self.bit_offset)
+                    .wrapping_shl((num_bits - self.bit_offset) as u32);
+            }
         }
 
         // TODO: better to avoid copying here
@@ -398,22 +397,7 @@ impl BitReader {
     ///
     /// Returns `false` if there are no more values to skip, `true` otherwise.
     pub fn skip_value(&mut self, num_bits: usize) -> bool {
-        assert!(num_bits <= 64);
-
-        if self.byte_offset * 8 + self.bit_offset + num_bits > self.total_bytes * 8 {
-            return false;
-        }
-
-        self.bit_offset += num_bits;
-
-        if self.bit_offset >= 64 {
-            self.byte_offset += 8;
-            self.bit_offset -= 64;
-
-            self.reload_buffer_values();
-        }
-
-        true
+        self.skip(1, num_bits) == 1
     }
 
     /// Read multiple values from their packed representation where each element is represented
@@ -429,7 +413,7 @@ impl BitReader {
 
         let mut values_to_read = batch.len();
         let needed_bits = num_bits * values_to_read;
-        let remaining_bits = (self.total_bytes - self.byte_offset) * 8 - self.bit_offset;
+        let remaining_bits = (self.buffer.len() - self.byte_offset) * 8 - self.bit_offset;
         if remaining_bits < needed_bits {
             values_to_read = remaining_bits / num_bits;
         }
@@ -536,8 +520,6 @@ impl BitReader {
             }
         }
 
-        self.reload_buffer_values();
-
         // Read any trailing values
         while i < values_to_read {
             let value = self
@@ -556,37 +538,25 @@ impl BitReader {
     pub fn skip(&mut self, num_values: usize, num_bits: usize) -> usize {
         assert!(num_bits <= 64);
 
-        let mut num_values = num_values;
         let needed_bits = num_bits * num_values;
-        let remaining_bits = (self.total_bytes - self.byte_offset) * 8 - self.bit_offset;
-        if remaining_bits < needed_bits {
-            num_values = remaining_bits / num_bits;
-        }
+        let remaining_bits = (self.buffer.len() - self.byte_offset) * 8 - self.bit_offset;
 
-        let mut values_skipped = 0;
+        let values_to_read = match remaining_bits < needed_bits {
+            true => remaining_bits / num_bits,
+            false => num_values,
+        };
 
-        // First align bit offset to byte offset
+        let end_bit_offset =
+            self.byte_offset * 8 + values_to_read * num_bits + self.bit_offset;
+
+        self.byte_offset = end_bit_offset / 8;
+        self.bit_offset = end_bit_offset % 8;
+
         if self.bit_offset != 0 {
-            while values_skipped < num_values && self.bit_offset != 0 {
-                self.skip_value(num_bits);
-                values_skipped += 1;
-            }
+            self.load_buffer()
         }
 
-        while num_values - values_skipped >= 32 {
-            self.byte_offset += 4 * num_bits;
-            values_skipped += 32;
-        }
-
-        assert!(num_values - values_skipped < 32);
-
-        self.reload_buffer_values();
-        while values_skipped < num_values {
-            self.skip_value(num_bits);
-            values_skipped += 1;
-        }
-
-        num_values
+        values_to_read
     }
 
     /// Reads up to `num_bytes` to `buf` returning the number of bytes read
@@ -596,7 +566,7 @@ impl BitReader {
         num_bytes: usize,
     ) -> usize {
         // Align to byte offset
-        self.byte_offset += ceil(self.bit_offset as i64, 8) as usize;
+        self.byte_offset = self.get_byte_offset();
         self.bit_offset = 0;
 
         let src = &self.buffer.data()[self.byte_offset..];
@@ -604,7 +574,6 @@ impl BitReader {
         buf.extend_from_slice(&src[..to_read]);
 
         self.byte_offset += to_read;
-        self.reload_buffer_values();
 
         to_read
     }
@@ -617,19 +586,17 @@ impl BitReader {
     /// Returns `Some` if there's enough bytes left to form a value of `T`.
     /// Otherwise `None`.
     pub fn get_aligned<T: FromBytes>(&mut self, num_bytes: usize) -> Option<T> {
-        let bytes_read = ceil(self.bit_offset, 8);
-        if self.byte_offset + bytes_read + num_bytes > self.total_bytes {
+        self.byte_offset = self.get_byte_offset();
+        self.bit_offset = 0;
+
+        if self.byte_offset + num_bytes > self.buffer.len() {
             return None;
         }
 
         // Advance byte_offset to next unread byte and read num_bytes
-        self.byte_offset += bytes_read;
         let v = read_num_bytes!(T, num_bytes, self.buffer.data()[self.byte_offset..]);
         self.byte_offset += num_bytes;
 
-        // Reset buffered_values
-        self.bit_offset = 0;
-        self.reload_buffer_values();
         Some(v)
     }
 
@@ -672,8 +639,10 @@ impl BitReader {
         })
     }
 
-    fn reload_buffer_values(&mut self) {
-        let bytes_to_read = cmp::min(self.total_bytes - self.byte_offset, 8);
+    /// Populates `self.buffered_values`
+    #[inline]
+    fn load_buffer(&mut self) {
+        let bytes_to_read = cmp::min(self.buffer.len() - self.byte_offset, 8);
         self.buffered_values =
             read_num_bytes!(u64, bytes_to_read, self.buffer.data()[self.byte_offset..]);
     }