# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

This makes a couple of changes:

```
arrow_array_reader/Int32Array/binary packed, mandatory, no NULLs                                                                             
                        time:   [26.773 us 26.778 us 26.784 us]
                        change: [-12.530% -12.270% -11.986%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low severe
  3 (3.00%) high mild
  2 (2.00%) high severe
arrow_array_reader/Int32Array/binary packed, optional, no NULLs                                                                             
                        time:   [43.490 us 43.501 us 43.513 us]
                        change: [-8.0569% -7.7694% -7.4579%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe
arrow_array_reader/Int32Array/binary packed, optional, half NULLs                                                                             
                        time:   [50.091 us 50.115 us 50.142 us]
                        change: [-3.0207% -2.7299% -2.4358%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
arrow_array_reader/Int64Array/binary packed, mandatory, no NULLs                                                                             
                        time:   [23.427 us 23.432 us 23.436 us]
                        change: [-7.3450% -7.2770% -7.2229%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high severe
arrow_array_reader/Int64Array/binary packed, optional, no NULLs                                                                             
                        time:   [40.159 us 40.169 us 40.179 us]
                        change: [-4.6524% -4.3771% -4.1016%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
arrow_array_reader/Int64Array/binary packed, optional, half NULLs                                                                             
                        time:   [49.294 us 49.312 us 49.330 us]
                        change: [-1.3314% -1.1008% -0.9611%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

arrow_array_reader/StringArray/plain encoded, mandatory, no NULLs                                                                            
                        time:   [171.83 us 171.96 us 172.09 us]
                        change: [-3.5398% -3.2856% -3.0931%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
arrow_array_reader/StringArray/plain encoded, optional, no NULLs                                                                            
                        time:   [189.01 us 189.08 us 189.16 us]
                        change: [-3.5964% -3.4081% -3.2798%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
arrow_array_reader/StringArray/plain encoded, optional, half NULLs                                                                            
                        time:   [211.78 us 211.91 us 212.03 us]
                        change: [-4.9192% -4.6823% -4.5268%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
arrow_array_reader/StringArray/dictionary encoded, mandatory, no NULLs                                                                            
                        time:   [134.94 us 135.22 us 135.51 us]
                        change: [-1.3736% -1.0797% -0.8383%] (p = 0.00 < 0.05)
                        Change within noise threshold.
arrow_array_reader/StringArray/dictionary encoded, optional, no NULLs                                                                            
                        time:   [150.57 us 150.81 us 151.06 us]
                        change: [-1.1385% -0.8209% -0.5214%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
arrow_array_reader/StringArray/dictionary encoded, optional, half NULLs                                                                            
                        time:   [188.25 us 188.37 us 188.49 us]
                        change: [-5.6464% -5.3275% -5.0459%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

arrow_array_reader/StringDictionary/dictionary encoded, mandatory, no NULLs                                                                             
                        time:   [24.404 us 24.413 us 24.423 us]
                        change: [-1.8395% -1.5543% -1.2527%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
arrow_array_reader/StringDictionary/dictionary encoded, optional, no NULLs                                                                             
                        time:   [41.200 us 41.229 us 41.271 us]
                        change: [-1.9306% -1.5861% -1.2388%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) high mild
  6 (6.00%) high severe
arrow_array_reader/StringDictionary/dictionary encoded, optional, half NULLs                                                                             
                        time:   [48.046 us 48.075 us 48.104 us]
                        change: [-7.1111% -6.7512% -6.3610%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe
```

# What changes are included in this PR?

1. Removes `total_bytes` making it easier for LLVM to elide bounds checks
2. Lazily populates buffered_values when actually needed
3. Simplifies the skip implementation

The second of these is particularly impactful for DeltaBitPackedDecoder, where the miniblocks are in chunks of 32 and therefore never actually fall out of byte alignment.

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?

No

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->